+++ neubird/aap/tests/unit/conftest.py
@@ -2,7 +2,7 @@
 import json
 import pytest
 from ansible.module_utils import basic
-from ansible.module_utils._text import to_bytes
+from ansible.module_utils.common.text.converters import to_bytes
 
 
 def set_module_args(args):